Add a getTape method to Tape class
A getTape() method will allow users to look at the tape. You should be
surprised I haven't implemented this yet. Right now, you have to
"delete" the integer array that is returned from this method, as it is
allocated on the heap. Speaking of documentation... I need to get a
leg up on that...