Create TestBed class for stringing test suites
A larger class is needed for containing multiple test suites to run
them all at once and have them in a central registry. (Also insert an
assert function. I don't necessarily think I like the assert function
to be in the TestSuite class, but I can't think of a more logical place
for it.)