Complicate by adding Case class
In order to keep good track of the tests that fail, it will be
necessary for the user of this test framework to provide a
description for each the test "Case", the assert, the test Suite,
and possibly (though not in this commit) the test Bed. In order to get
this kind of functionality, a Case class seems necessary. Hopefully
I can remember where I was going with all of this tomorrow. This is
supposed to streamline the user experience somehow.