Add message handling to framework
The framework needs a way of handling failed tests so that the whole
flow of the test run isn't stopped as soon as a single assert fails.
To do this, add more advanced messaging allowing the user to tack a
description to each assert, test case, and test suite to find where a
particular test failed.

Also add output functionality. When the tests run in a Suite, pass them
an ostream& to print output to. This will allow a user to see whether
the tests he tagged with his descriptions passed or failed.

Possibly in the future (not in this commit), add an option for the user
to print out assert descriptions even if they pass. I'll leave that as
a TODO for now.